[최병호] 코드 분리
