+++ vanilla-movie/src/main.ts
@@ -9,6 +9,8 @@ import { FineDustService } from "./api/services/PData/fineDust.service.ts";
 import { LectureListService } from "./api/services/PData/lectureList.service.ts";
 import { ReviewListService } from "./api/services/tmdb/reviewList.service.ts";
 import { PDataService } from "./api/services/PData/index.service.ts";
+import { BuchonSubwayPassengerService } from "./api/services/PData/bucheonSubwayPassenger.service.ts";
+import { CoronaMediaContentChangeService } from "./api/services/PData/coronaMediaContentChange.service.ts";
 
 const movieListService = new MovieListService();
 const weatherService = new WeatherService();
@@ -17,8 +19,18 @@ const fineDustService = new FineDustService();
 const lectureListService = new LectureListService();
 const reviewListService = new ReviewListService();
 const pDataService = new PDataService();
+const coronaMediaContentChangeService = new CoronaMediaContentChangeService();
+const bucheonSubwayPassengerService = new BuchonSubwayPassengerService();
 
 (async () => {
+  // const corona =
+  //   await coronaMediaContentChangeService.getCoronaMediaContentChange({
+  //     params: {
+  //       page: 1,
+  //       perPage: 10,
+  //     },
+  //   });
+  // console.log(corona);
   // const data = await movieListService.getPopularMovies({
   //   params: {
   //     include_adult: false,
@@ -34,15 +46,14 @@ const pDataService = new PDataService();
   //     lon: 126.978,
   //   },
   // });
-  // const buchonSubwayPassengerData = await pDataService.getBucheonSubwayPassenger({
+  // const buchonSubwayPassengerData = await bucheonSubwayPassengerService.getBucheonSubwayPassenger({
   //   params: {
   //     page: 1,
   //     perPage: 10,
   //     returnType: "json",
   //   },
   // });
-  // console.log("buchonSubwayPassengerData", buchonSubwayPassengerData);
-  // console.log("buchonSubwayPassengerData", buchonSubwayPassengerData);
+  // console.log("buchonSubwayPassengerData", buchonSubwayPassengerData);\
   // await console.log("page", data.page);
   // const corona = await pDataService.getCoronaMediaContentChange({
   //   params: {
